requirements.txt
requests==2.28.1
python-dotenv==0.21.0
gunicorn==20.1.0
Flask==2.3.3
Werkzeug==2.3.7
facebook-sdk==3.1.0
google-ads==22.1.0
twitter-ads==11.0.0